team_reflect_changes_from_outside_of_NM: fix the test
Spawn the teamd daemon for team0 creation. It cannot just be created with
rtnetlink, since subsequent teamdctl would have noone to talk to.
Also, the link up of dummy device is unnecessary; NM would then assume a
connection and then we wouldn't be able to test that NM assumes the
connection upon team enslavement.

Change-Id: Ibb63341b0456fd7fef2ab79af57a41414368faec