ipv6_routes_set_basic_route: don't let the connections autoconnect
We bring them up anyways.

Thing is that we modify method and if the connection was configuring a
link-local address at the time it would easily hit an assertion.

This is a known bug thats' being fixed, but let's not let it break down
testing of a different feature for now:

https://bugzilla.gnome.org/show_bug.cgi?id=724041

Change-Id: Ia228da3c9c559acf9ac970d5c635cf9a6ffc055e