avoid needlesssly setting ipv6.method to ignore
Nothing really wrong with that, but it likes to trigger bug rh #724041 more
often than we'd like it to. We're eventually going to fix the bug, but it's
not likely to happen too often in real use cases so let's just not let it
mess with testing of unrelated features.

It's really rather okay the leave it with auto setting when the connection
will happily finish activation even when there's no global IPv6 addresses.

Change-Id: Iec155e1d64943b7bb4d5bf73e4b36c7e5b3e58f7