run_once_new_connection: be a bit more patient when waiting for the helper
The helper is started only when the connection is fully activated.
NM was changed to apply the IPv6 address as soon as it receives a RA, but that
doesn't necessarily mean the IPv6 configuration is done. Give it some
more time.

Also, give the addresses some time too to make it more robust. The packet
can get lost and it takes time to retransmit them.

Change-Id: Ia6fdc7944340b9586d13e76307d81d6f820f5667