environment: hack around NetworkManager-config-server trickery
We install the package from a repository and it could be the wrong
version. The 1.2 package places the configuration in /usr/lib while the
1.0 package has it in /etc. 1.0 doesn't know about /usr/lib.

Shit.

The files are otherwise the same: Let's just hack around it by shuffling
the files around. One day we may forget 1.0 and just mask the file by
touching a file in /etc.

Change-Id: Ia6931885addcf941c85bf0d9c9e13f6b523c32fd