nmtui: ipv6: don't assume that connection is up when addresses appear
NM 1.3 waits that IPv6 addresses become non-tentative before signaling
the connection as up, because tentative addresses are not usable yet
by applications (see rh #1243958).

Thus, the presence of the configured address does not imply that the
connection is up. Instead, check first that the connection is up and
then that the addresses are configured. Also, disable IPv4 because
otherwise we can't be sure whether connection went up due to IPv4 or
IPv6.

Change-Id: I806b7eee9d580d10c4b6c8d52177f14236af6bdc