bond: differentiate between interface and link state
Previously we read the value of "MII Status" line in
/proc/net/bonding/$ifname to determine if a bond interface was up. But
that value represents the *link* state and in general, for bonds
without slaves, the link state is down. But due to a kernel bug the
initial value of the property was up, and tests depended on that
value.

This changed with new NM because we write the 'min_links' option when
initializing the bond and this triggers kernel to update the link
state to the correct (down) value.

Fix the tests to read the correct state:

 - for bonds which have been brought up but don't have slaves, check
   the interface state ('ip addr show $ifname up' output)

 - for bonds up with slaves, still check the link state after "MII
   Status" string

Change-Id: I2d0f8b50b144d1b0ce022c87bdb03b56d01e414a