general: explicitly activate connection in @set_device_unmanaged
The auto-activation is not synchronous to "nmcli add" command and thus
the dhclient process is not guaranteed to be present when the command
terminates. Also, DHCP may take some time to complete and so the
address can be missing when we unmanage the device.

To fix this, explicitly activate the connection.

Change-Id: I41ef2a9d5697ea0eea817c2bbd634a6769bd8fa6